# perf: improve performance by adding `inline` and `cold` attributes, and slice access over index access