--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "xsum"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 rust-version = "1.69"
 license = "MIT"