- Improve performance by adding `inline` and `cold` attributes, and slice access over index access
- Improve benchmark code
- Improve test cases
- Update dev dependencies

### Benchmark comparison

Compared the latest main at this moment (https://github.com/Gumichocopengin8/xsum.rs/commit/3add6bec24bba69c999bb2d21d65b730457adbdd) and this PR change.

Overall, the benchmarks are improved or the same. `xsum/xsumauto add_list/50000`, `xsum/xsumlarge add/100`, and `xsum/xsumauto add_list/100` are regressed up to `1.7%` except `xsum/xsumlarge add/100` (`4.9%` regression). However, `xsum/xsumlarge add/100` shouldn't be used with 100 data size in practice, so it's improved overall.

```bash
$ cargo bench xsum
   Compiling xsum v0.1.2 (/xsum)
    Finished `bench` profile [optimized] target(s) in 15.88s
     Running unittests src/lib.rs (target/release/deps/xsum-ad971a16d0e9e0be)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running benches/bench_xsum_sum.rs (target/release/deps/bench_xsum_sum-14151473bf60571f)
Gnuplot not found, using plotters backend
xsum/xsumsmall add_list/1000
                        time:   [2.1481 µs 2.1498 µs 2.1516 µs]
                        thrpt:  [464.78 Melem/s 465.15 Melem/s 465.54 Melem/s]
                 change:
                        time:   [+0.9358% +1.0160% +1.0989%] (p = 0.00 < 0.05)
                        thrpt:  [−1.0869% −1.0058% −0.9272%]
                        Change within noise threshold.
xsum/xsumsmall add/1000 time:   [2.2511 µs 2.2529 µs 2.2552 µs]
                        thrpt:  [443.42 Melem/s 443.87 Melem/s 444.23 Melem/s]
                 change:
                        time:   [−35.569% −35.514% −35.463%] (p = 0.00 < 0.05)
                        thrpt:  [+54.949% +55.073% +55.205%]
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
xsum/xsumlarge add_list/1000
                        time:   [3.3660 µs 3.3922 µs 3.4191 µs]
                        thrpt:  [292.47 Melem/s 294.80 Melem/s 297.09 Melem/s]
                 change:
                        time:   [−0.7035% +0.7879% +2.4246%] (p = 0.31 > 0.05)
                        thrpt:  [−2.3672% −0.7817% +0.7084%]
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild
xsum/xsumlarge add/1000 time:   [3.4146 µs 3.4518 µs 3.4909 µs]
                        thrpt:  [286.46 Melem/s 289.70 Melem/s 292.86 Melem/s]
                 change:
                        time:   [−4.7727% −3.6108% −2.4862%] (p = 0.00 < 0.05)
                        thrpt:  [+2.5496% +3.7460% +5.0119%]
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  5 (5.00%) low severe
  4 (4.00%) high mild
  8 (8.00%) high severe
xsum/xsumauto add_list/1000
                        time:   [2.1507 µs 2.1526 µs 2.1550 µs]
                        thrpt:  [464.03 Melem/s 464.55 Melem/s 464.97 Melem/s]
                 change:
                        time:   [+0.9941% +1.1309% +1.2684%] (p = 0.00 < 0.05)
                        thrpt:  [−1.2525% −1.1182% −0.9843%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
xsum/xsumauto add/1000  time:   [2.3777 µs 2.3821 µs 2.3866 µs]
                        thrpt:  [419.01 Melem/s 419.79 Melem/s 420.57 Melem/s]
                 change:
                        time:   [−31.861% −31.749% −31.643%] (p = 0.00 < 0.05)
                        thrpt:  [+46.291% +46.517% +46.759%]
                        Performance has improved.
xsum/xsumsmall add_list/5000
                        time:   [10.456 µs 10.462 µs 10.470 µs]
                        thrpt:  [477.57 Melem/s 477.92 Melem/s 478.19 Melem/s]
                 change:
                        time:   [+0.4090% +0.5760% +0.7350%] (p = 0.00 < 0.05)
                        thrpt:  [−0.7297% −0.5727% −0.4074%]
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
xsum/xsumsmall add/5000 time:   [10.997 µs 11.004 µs 11.013 µs]
                        thrpt:  [454.01 Melem/s 454.40 Melem/s 454.67 Melem/s]
                 change:
                        time:   [−36.105% −35.994% −35.851%] (p = 0.00 < 0.05)
                        thrpt:  [+55.886% +56.235% +56.507%]
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) high mild
  10 (10.00%) high severe
xsum/xsumlarge add_list/5000
                        time:   [11.861 µs 11.907 µs 11.949 µs]
                        thrpt:  [418.45 Melem/s 419.90 Melem/s 421.53 Melem/s]
                 change:
                        time:   [−2.1675% −1.6849% −1.1964%] (p = 0.00 < 0.05)
                        thrpt:  [+1.2109% +1.7138% +2.2155%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low mild
xsum/xsumlarge add/5000 time:   [12.028 µs 12.075 µs 12.122 µs]
                        thrpt:  [412.48 Melem/s 414.09 Melem/s 415.70 Melem/s]
                 change:
                        time:   [−1.4394% −1.0794% −0.7038%] (p = 0.00 < 0.05)
                        thrpt:  [+0.7088% +1.0912% +1.4605%]
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
xsum/xsumauto add_list/5000
                        time:   [11.801 µs 11.845 µs 11.895 µs]
                        thrpt:  [420.35 Melem/s 422.12 Melem/s 423.70 Melem/s]
                 change:
                        time:   [−0.8175% −0.4602% −0.1389%] (p = 0.01 < 0.05)
                        thrpt:  [+0.1391% +0.4624% +0.8242%]
                        Change within noise threshold.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  6 (6.00%) high mild
  5 (5.00%) high severe
xsum/xsumauto add/5000  time:   [12.149 µs 12.177 µs 12.207 µs]
                        thrpt:  [409.61 Melem/s 410.62 Melem/s 411.56 Melem/s]
                 change:
                        time:   [−8.3213% −8.0127% −7.7139%] (p = 0.00 < 0.05)
                        thrpt:  [+8.3587% +8.7106% +9.0766%]
                        Performance has improved.
xsum/xsumsmall add_list/20000
                        time:   [41.606 µs 41.622 µs 41.638 µs]
                        thrpt:  [480.33 Melem/s 480.51 Melem/s 480.70 Melem/s]
                 change:
                        time:   [+0.1899% +0.3364% +0.4574%] (p = 0.00 < 0.05)
                        thrpt:  [−0.4553% −0.3352% −0.1896%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
xsum/xsumsmall add/20000
                        time:   [45.224 µs 45.296 µs 45.370 µs]
                        thrpt:  [440.82 Melem/s 441.54 Melem/s 442.24 Melem/s]
                 change:
                        time:   [−34.251% −34.151% −34.044%] (p = 0.00 < 0.05)
                        thrpt:  [+51.617% +51.862% +52.095%]
                        Performance has improved.
xsum/xsumlarge add_list/20000
                        time:   [43.136 µs 43.189 µs 43.246 µs]
                        thrpt:  [462.47 Melem/s 463.08 Melem/s 463.65 Melem/s]
                 change:
                        time:   [−0.6889% −0.5748% −0.4557%] (p = 0.00 < 0.05)
                        thrpt:  [+0.4578% +0.5781% +0.6936%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
xsum/xsumlarge add/20000
                        time:   [44.031 µs 44.112 µs 44.201 µs]
                        thrpt:  [452.48 Melem/s 453.39 Melem/s 454.23 Melem/s]
                 change:
                        time:   [−0.5813% −0.4081% −0.2473%] (p = 0.00 < 0.05)
                        thrpt:  [+0.2479% +0.4097% +0.5847%]
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) low mild
  5 (5.00%) high mild
  1 (1.00%) high severe
xsum/xsumauto add_list/20000
                        time:   [43.136 µs 43.180 µs 43.223 µs]
                        thrpt:  [462.72 Melem/s 463.18 Melem/s 463.65 Melem/s]
                 change:
                        time:   [+0.2128% +0.3290% +0.4533%] (p = 0.00 < 0.05)
                        thrpt:  [−0.4513% −0.3279% −0.2124%]
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  2 (2.00%) low severe
  9 (9.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe
xsum/xsumauto add/20000 time:   [43.963 µs 44.010 µs 44.064 µs]
                        thrpt:  [453.88 Melem/s 454.44 Melem/s 454.93 Melem/s]
                 change:
                        time:   [−3.2640% −3.1721% −3.0736%] (p = 0.00 < 0.05)
                        thrpt:  [+3.1711% +3.2761% +3.3741%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) low mild
  2 (2.00%) high severe
xsum/xsumsmall add_list/100000
                        time:   [207.22 µs 207.44 µs 207.74 µs]
                        thrpt:  [481.38 Melem/s 482.06 Melem/s 482.58 Melem/s]
                 change:
                        time:   [+0.3141% +0.4291% +0.5450%] (p = 0.00 < 0.05)
                        thrpt:  [−0.5421% −0.4272% −0.3131%]
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
xsum/xsumsmall add/100000
                        time:   [220.18 µs 220.74 µs 221.40 µs]
                        thrpt:  [451.67 Melem/s 453.01 Melem/s 454.17 Melem/s]
                 change:
                        time:   [−35.733% −35.608% −35.487%] (p = 0.00 < 0.05)
                        thrpt:  [+55.007% +55.298% +55.601%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
xsum/xsumlarge add_list/100000
                        time:   [208.95 µs 209.18 µs 209.46 µs]
                        thrpt:  [477.41 Melem/s 478.05 Melem/s 478.59 Melem/s]
                 change:
                        time:   [−1.1115% −1.0158% −0.9200%] (p = 0.00 < 0.05)
                        thrpt:  [+0.9285% +1.0262% +1.1240%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
xsum/xsumlarge add/100000
                        time:   [212.20 µs 212.26 µs 212.32 µs]
                        thrpt:  [470.98 Melem/s 471.11 Melem/s 471.25 Melem/s]
                 change:
                        time:   [−0.2580% −0.1794% −0.0931%] (p = 0.00 < 0.05)
                        thrpt:  [+0.0932% +0.1797% +0.2587%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
xsum/xsumauto add_list/100000
                        time:   [208.57 µs 208.61 µs 208.65 µs]
                        thrpt:  [479.27 Melem/s 479.36 Melem/s 479.46 Melem/s]
                 change:
                        time:   [−0.5641% −0.2463% +0.0092%] (p = 0.10 > 0.05)
                        thrpt:  [−0.0092% +0.2469% +0.5673%]
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
xsum/xsumauto add/100000
                        time:   [212.44 µs 212.60 µs 212.78 µs]
                        thrpt:  [469.98 Melem/s 470.38 Melem/s 470.71 Melem/s]
                 change:
                        time:   [−1.0739% −1.0018% −0.9252%] (p = 0.00 < 0.05)
                        thrpt:  [+0.9339% +1.0120% +1.0856%]
                        Change within noise threshold.
Found 16 outliers among 100 measurements (16.00%)
  6 (6.00%) high mild
  10 (10.00%) high severe
xsum/xsumsmall add_list/10000
                        time:   [20.789 µs 20.798 µs 20.808 µs]
                        thrpt:  [480.60 Melem/s 480.80 Melem/s 481.02 Melem/s]
                 change:
                        time:   [+0.2333% +0.3215% +0.3944%] (p = 0.00 < 0.05)
                        thrpt:  [−0.3928% −0.3205% −0.2328%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
xsum/xsumsmall add/10000
                        time:   [21.871 µs 21.876 µs 21.882 µs]
                        thrpt:  [457.00 Melem/s 457.12 Melem/s 457.22 Melem/s]
                 change:
                        time:   [−36.359% −36.316% −36.269%] (p = 0.00 < 0.05)
                        thrpt:  [+56.909% +57.025% +57.132%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
xsum/xsumlarge add_list/10000
                        time:   [22.256 µs 22.285 µs 22.318 µs]
                        thrpt:  [448.08 Melem/s 448.74 Melem/s 449.31 Melem/s]
                 change:
                        time:   [−0.4338% −0.3161% −0.1937%] (p = 0.00 < 0.05)
                        thrpt:  [+0.1941% +0.3171% +0.4357%]
                        Change within noise threshold.
xsum/xsumlarge add/10000
                        time:   [22.621 µs 22.679 µs 22.744 µs]
                        thrpt:  [439.67 Melem/s 440.94 Melem/s 442.08 Melem/s]
                 change:
                        time:   [−0.8060% −0.4921% −0.2085%] (p = 0.00 < 0.05)
                        thrpt:  [+0.2089% +0.4945% +0.8126%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high severe
xsum/xsumauto add_list/10000
                        time:   [22.139 µs 22.183 µs 22.225 µs]
                        thrpt:  [449.95 Melem/s 450.79 Melem/s 451.68 Melem/s]
                 change:
                        time:   [−0.0303% +0.1665% +0.3513%] (p = 0.09 > 0.05)
                        thrpt:  [−0.3501% −0.1663% +0.0303%]
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
xsum/xsumauto add/10000 time:   [22.826 µs 22.878 µs 22.922 µs]
                        thrpt:  [436.27 Melem/s 437.10 Melem/s 438.10 Melem/s]
                 change:
                        time:   [−4.8648% −4.6763% −4.4712%] (p = 0.00 < 0.05)
                        thrpt:  [+4.6805% +4.9057% +5.1136%]
                        Performance has improved.
xsum/xsumsmall add_list/50000
                        time:   [103.63 µs 103.75 µs 103.93 µs]
                        thrpt:  [481.09 Melem/s 481.92 Melem/s 482.49 Melem/s]
                 change:
                        time:   [−0.0564% +0.1028% +0.2458%] (p = 0.19 > 0.05)
                        thrpt:  [−0.2452% −0.1027% +0.0564%]
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
xsum/xsumsmall add/50000
                        time:   [111.15 µs 111.36 µs 111.60 µs]
                        thrpt:  [448.01 Melem/s 449.00 Melem/s 449.85 Melem/s]
                 change:
                        time:   [−35.617% −35.495% −35.372%] (p = 0.00 < 0.05)
                        thrpt:  [+54.732% +55.028% +55.321%]
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  12 (12.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
xsum/xsumlarge add_list/50000
                        time:   [105.15 µs 105.18 µs 105.21 µs]
                        thrpt:  [475.23 Melem/s 475.37 Melem/s 475.53 Melem/s]
                 change:
                        time:   [−0.8301% −0.7486% −0.6796%] (p = 0.00 < 0.05)
                        thrpt:  [+0.6843% +0.7542% +0.8371%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
xsum/xsumlarge add/50000
                        time:   [107.03 µs 107.18 µs 107.38 µs]
                        thrpt:  [465.62 Melem/s 466.52 Melem/s 467.14 Melem/s]
                 change:
                        time:   [−0.7241% −0.5773% −0.4315%] (p = 0.00 < 0.05)
                        thrpt:  [+0.4333% +0.5806% +0.7294%]
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
xsum/xsumauto add_list/50000
                        time:   [106.33 µs 106.72 µs 107.12 µs]
                        thrpt:  [466.78 Melem/s 468.53 Melem/s 470.24 Melem/s]
                 change:
                        time:   [+1.3924% +1.7108% +2.0061%] (p = 0.00 < 0.05)
                        thrpt:  [−1.9666% −1.6820% −1.3733%]
                        Performance has regressed.
xsum/xsumauto add/50000 time:   [107.12 µs 107.15 µs 107.18 µs]
                        thrpt:  [466.51 Melem/s 466.64 Melem/s 466.77 Melem/s]
                 change:
                        time:   [−1.4833% −1.4322% −1.3782%] (p = 0.00 < 0.05)
                        thrpt:  [+1.3974% +1.4530% +1.5057%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
xsum/xsumsmall add_list/10
                        time:   [94.991 ns 95.157 ns 95.326 ns]
                        thrpt:  [104.90 Melem/s 105.09 Melem/s 105.27 Melem/s]
                 change:
                        time:   [−4.8487% −4.3729% −3.8866%] (p = 0.00 < 0.05)
                        thrpt:  [+4.0438% +4.5728% +5.0958%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
xsum/xsumsmall add/10   time:   [97.947 ns 98.112 ns 98.289 ns]
                        thrpt:  [101.74 Melem/s 101.92 Melem/s 102.10 Melem/s]
                 change:
                        time:   [−7.5145% −7.1245% −6.7338%] (p = 0.00 < 0.05)
                        thrpt:  [+7.2200% +7.6710% +8.1251%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
xsum/xsumlarge add_list/10
                        time:   [1.1931 µs 1.2196 µs 1.2448 µs]
                        thrpt:  [8.0335 Melem/s 8.1995 Melem/s 8.3818 Melem/s]
                 change:
                        time:   [−6.2744% −3.4061% −0.5441%] (p = 0.02 < 0.05)
                        thrpt:  [+0.5471% +3.5262% +6.6945%]
                        Change within noise threshold.
Found 21 outliers among 100 measurements (21.00%)
  10 (10.00%) low severe
  1 (1.00%) low mild
  8 (8.00%) high mild
  2 (2.00%) high severe
xsum/xsumlarge add/10   time:   [909.40 ns 947.04 ns 993.09 ns]
                        thrpt:  [10.070 Melem/s 10.559 Melem/s 10.996 Melem/s]
                 change:
                        time:   [−11.610% −7.9471% −3.9193%] (p = 0.00 < 0.05)
                        thrpt:  [+4.0792% +8.6332% +13.135%]
                        Performance has improved.
xsum/xsumauto add_list/10
                        time:   [95.757 ns 95.940 ns 96.152 ns]
…